nvim: Move config under `jagd` namespace
