+++ nvim/.config/nvim/init.lua
@@ -1,6 +1,6 @@
-require("config.leader")
-require("config.lazy")
-require("config.options")
-require("config.autocmds")
-require("config.keymaps")
-require("config.colorscheme")
+require("jagd.config.leader")
+require("jagd.config.lazy")
+require("jagd.config.options")
+require("jagd.config.autocmds")
+require("jagd.config.keymaps")
+require("jagd.config.colorscheme")
