+++ nvim/.config/nvim/lua/jagd/config/autocmds.lua
@@ -1,5 +1,5 @@
 local autocmd = vim.api.nvim_create_autocmd
-local augroup = require("utils").augroup
+local augroup = require("jagd.utils").augroup
 
 -- Only show cursor line in current split
 local cursorline_group = augroup("cursorline")
