+++ nvim/.config/nvim/lua/jagd/config/options.lua
