+++ nvim/.config/nvim/lua/jagd/plugins/blink.lua
