+++ nvim/.config/nvim/lua/jagd/plugins/lualine.lua
