+++ nvim/.config/nvim/lua/jagd/plugins/mason.lua
