+++ nvim/.config/nvim/lua/jagd/plugins/obsidian.lua
