+++ nvim/.config/nvim/lua/jagd/plugins/trouble.lua
