+++ nvim/.config/nvim/lua/jagd/plugins/vim-sleuth.lua
